relationship edit drawer updated to keep the drawer behavior, tsx updated
