+++ opencti-platform/opencti-front/src/private/components/common/stix_core_relationships/StixCoreRelationshipPopover.jsx
@@ -141,7 +141,7 @@ class StixCoreRelationshipPopover extends Component {
             {t('Delete')}
           </MenuItem>
         </Menu>
-        {stixCoreRelationshipId && this.state.displayUpdate && (
+        {stixCoreRelationshipId && (
           <StixCoreRelationshipEdition
             variant="noGraph"
             stixCoreRelationshipId={stixCoreRelationshipId}
